+++ precompile/utils/src/lib.rs
@@ -41,8 +41,6 @@ pub use data::{Address, Bytes, EvmData, EvmDataReader, EvmDataWriter};
 pub use precompile_utils_macro::{generate_function_selector, keccak256};
 
 #[cfg(feature = "testing")]
-pub mod mock;
-#[cfg(feature = "testing")]
 pub mod testing;
 #[cfg(test)]
 mod tests;
